Item9415: Make Var* topic parents consistent
We already know that a Var* topic is a Macro topic; so it's pointless
for plugin-provided MACROs to be parented under the Macro topic.

Better to parent them to the plugin's system topic, so that users can discover
what macros are provided by their plugin (assuming a site skin that offers
hierarchical browsing, admittedly we don't have this foswiki core yet).

Idea came from the fact that JQueryPlugin's Var* topics are parented this way.

git-svn-id: http://svn.foswiki.org/branches/Release01x01/TablePlugin@9224 0b4bb1d4-4e5a-0410-9cc4-b2b747904278